move filter.py out of prep folder
